Rename filename to integer_round.js
