+++ benchmarks/integer_round.js
